Add sw-precache custom typing
