+++ tsconfig.json
@@ -28,6 +28,7 @@
     "custom_typings/node.d.ts",
     "custom_typings/plylog.d.ts",
     "custom_typings/slash.d.ts",
+    "custom_typings/sw-precache.d.ts",
     "custom_typings/vinyl-fs.d.ts",
     "src/analyzer.ts",
     "src/bundle.ts",
